Finished part one of the project
Created project1. Inside the project I made packages for testing,
models, viewing the ui, and made class main. In model package, there is
class FoodItem that holds the info of one item, and Pantry which holds 5
items in an array. Test package has two packages for testing Fooditem and Pantry.
test/fooditem holds TestConstructor and TestSetQuantity. test/pantry holds
TestPantry. view package holds the codebehind class aka MainWindow and the FXML file.
All of this is used to make a UI that adds items to a pantry that has a type and a certain amount.