update project2-3B recipe load
I only updated half of it so far to get help on it.
Created a static method in the utility class to convert a list of names to a string.
created a way to load a file of with names of recipes and test for it however it is broken and unfinished.